refactor: reorganize ports and update import paths across multiple files
