+++ cmd/server/main.go
@@ -10,12 +10,12 @@ import (
 	"os/signal"
 	"pinstack-auth-service/config"
 	token_service "pinstack-auth-service/internal/application/service"
-	authManager "pinstack-auth-service/internal/infrastructure/auth"
-	user "pinstack-auth-service/internal/infrastructure/client/user"
-	delivery_grpc "pinstack-auth-service/internal/infrastructure/grpc"
-	auth_grpc "pinstack-auth-service/internal/infrastructure/grpc/auth"
+	delivery_grpc "pinstack-auth-service/internal/infrastructure/inbound/grpc"
+	auth_grpc "pinstack-auth-service/internal/infrastructure/inbound/grpc/auth"
 	infraLogger "pinstack-auth-service/internal/infrastructure/logger"
-	token_repository "pinstack-auth-service/internal/infrastructure/repository/postgres"
+	authManager "pinstack-auth-service/internal/infrastructure/outbound/auth"
+	user "pinstack-auth-service/internal/infrastructure/outbound/client/user"
+	token_repository "pinstack-auth-service/internal/infrastructure/outbound/repository/postgres"
 	"syscall"
 	"time"
