+++ internal/infrastructure/inbound/grpc/auth/api.go
@@ -1,7 +1,8 @@
 package auth_grpc
 
 import (
-	ports "pinstack-auth-service/internal/domain/ports"
+	input "pinstack-auth-service/internal/domain/ports/input"
+	ports "pinstack-auth-service/internal/domain/ports/output"
 
 	"github.com/go-playground/validator/v10"
 	pb "github.com/soloda1/pinstack-proto-definitions/gen/go/pinstack-proto-definitions/auth/v1"
@@ -11,11 +12,11 @@ var validate = validator.New()
 
 type AuthGRPCService struct {
 	pb.UnimplementedAuthServiceServer
-	tokenService ports.TokenService
+	tokenService input.TokenService
 	log          ports.Logger
 }
 
-func NewAuthGRPCService(tokenService ports.TokenService, log ports.Logger) *AuthGRPCService {
+func NewAuthGRPCService(tokenService input.TokenService, log ports.Logger) *AuthGRPCService {
 	return &AuthGRPCService{
 		tokenService: tokenService,
 		log:          log,
